Addressed all comments
